Renamed components used only by Web Legacy to {name}Legacy
