+++ src/routes/PageDetailsLegacy/components/PageDetailsForm/styles.module.scss
