+++ src/routes/PageDetailsLegacy/styles.module.scss
