+++ src/routes/PaymentLegacy/components/PaymentFormLegacy/index.jsx
@@ -19,7 +19,7 @@ import styles from "./styles.module.scss";
 /**
  * Payment Form Page
  */
-const PaymentForm = ({ formData, setFormData }) => {
+const PaymentFormLegacy = ({ formData, setFormData }) => {
   const handleInputChange = (name, value) =>
     setFormData((formData) => ({ ...formData, [name]: value }));
 
@@ -161,12 +161,12 @@ const PaymentForm = ({ formData, setFormData }) => {
   );
 };
 
-PaymentForm.defaultProps = {};
+PaymentFormLegacy.defaultProps = {};
 
-PaymentForm.propTypes = {
+PaymentFormLegacy.propTypes = {
   formData: PT.shape(),
   onFormUpdate: PT.func,
   setFormData: PT.func.isRequired,
 };
 
-export default PaymentForm;
+export default PaymentFormLegacy;
